get_token_user 모든 앱에서 적용
